pyproject.toml
[project]
name = "fast-soy-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "argon2-cffi>=25.1.0",
  "email-validator>=2.3.0",
  "fastapi>=0.121.2",
  "fastapi-cache2[redis]>=0.2.2",
  "fastapi-guard>=5.0.0",
  "httpx>=0.28.1",
  "loguru>=0.7.3",
  "orjson>=3.11.7",
  "passlib>=1.7.4",
  "pretty-errors>=1.2.25",
  "psutil>=7.2.2",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.13.1",
  "pyjwt>=2.12.1",
  "starlette==0.52.1",
  "tortoise-orm>=1.1.7",
  "uvicorn>=0.42.0",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.28.1"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.pyright]
include = ["app"]
exclude = [
  "web",
  ".venv",
  "**/.*",
  "**/__pycache__",
  "**/node_modules",
  "src/experimental",
  "src/typestubs",
]
pythonVersion = "3.12"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
  { root = "./tests/python_3_12", pythonVersion = "3.12", extraPaths = [
    "./",
  ] },
  { root = "./tests", extraPaths = [
    "./",
  ] },
  { root = "./app/system/schemas", reportIncompatibleVariableOverride = false, extraPaths = [
    "./",
  ] },
  { root = "./" },
]

typeCheckingMode = "standard"
disableBytesTypePromotions = true
